+++ packages/@react-spectrum/story-utils/src/index.ts
@@ -13,3 +13,4 @@
 /// <reference types="css-module-types" />
 
 export * from './ErrorBoundary';
+export * from './GeneratePowerset';
